textai-website-builder/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
openai==1.3.5
google-generativeai==0.2.2
anthropic==0.3.11
torch==2.1.0
pinecone-client==2.2.4
requests==2.31.0
python-dotenv==1.0.0
stripe==8.9.0